Large-file support for 64-bit systems
Apparently, only a few lines have to be changed in order to support
large(r) files on 64-bit machines.

This commit doesn't (yet) fix the issue on 32-bit machines (it also
doesn't test this explicitly). In comparison to #59, however, it uses
types that help get this to work on 32-bit machines as well, as it
doesn't use compiler-dependent types, but types that are known to be
large enough even there.

Closes #59.

CC #31.