Add support for SSL_CERT_FILE environment variable
This environment variable is supported by the CLI curl tool, too. It allows using a custom CA bundle with the tool.

Reasons to use this environment variable include the use of self-signed or otherwise untrusted certificates on the server or setting a CA bundle path when using the tool on a distro whose path differs from the one used on the build system.