Use FetchContent to include CPR
This also changes the logic to respect CPR's CPR_USE_SYSTEM_CURL. If that variable is not set, we configure CPR to use the system CURL if it is old enough to not cause issues and build its own otherwise. This allows for developing on recent systems but allows the easy use of a system libcurl when building in older distros for deployment.